# lib/api-spec/msp-openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title — it controls generated filenames (MspApi → msp-api.ts)
  title: MspApi
  version: 0.1.0
  description: MSP Portal API specification — covers authentication, MSP CRUD, customer management, and portal-facing endpoints.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /msp/customers/{id}:
    get:
      operationId: getMspCustomer
      tags: [msp-customers]
      summary: Get a single customer record
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MspCustomer"
        "404":
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MspErrorResponse"

    patch:
      operationId: updateMspCustomer
      tags: [msp-customers]
      summary: Update a customer record
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMspCustomerInput"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MspCustomer"
        "404":
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MspErrorResponse"

  # ── Signup / Tiers ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MspCustomer:
      type: object
      required: [id, companyName, status]
      properties:
        id:
          type: integer
        mspId:
          type: integer
        companyName:
          type: string
        domain:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, suspended, offboarding]
        contactName:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
        tenantId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    UpdateMspCustomerInput:
      type: object
      properties:
        companyName:
          type: string
          minLength: 1
        domain:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
          format: email
        status:
          type: string
          enum: [active, inactive, suspended]

    # ── Signup / Tiers ────────────────────────────────────────────────────────
...
    MspErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true